--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,6 @@
+golang-fifo.test
+out
+
 # IDE
 .idea
 .vscode