--- a/go.mod
+++ b/go.mod
@@ -2,7 +2,10 @@ module github.com/scalalang2/golang-fifo
 
 go 1.22
 
-require github.com/stretchr/testify v1.8.4
+require (
+	github.com/cockroachdb/swiss v0.0.0-20240303172742-c161743eb608
+	github.com/stretchr/testify v1.8.4
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect